requirements.txt
# Flask and extensions
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3

# Database
SQLAlchemy==2.0.23

# Password hashing
bcrypt==4.1.1

# AI Models - Sentence Transformers
sentence-transformers==2.2.2
torch

# OCR
pytesseract==0.3.10
Pillow==10.1.0

# Data processing
numpy==1.26.2
scikit-learn==1.3.2

# Chart.js integration (via CDN in templates)
# No additional Python packages needed

# Development
python-dotenv==1.0.0
Werkzeug==3.0.1
